backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

SQLAlchemy>=2.0.0
alembic>=1.13.0
pydantic-settings>=2.4.0
email-validator>=2.1.0

python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4

PyMySQL>=1.1.0

huggingface_hub>=0.24.0
httpx>=0.27.0

sse-starlette>=2.1.0

llama-cpp-python>=0.2.90
